pyproject.toml
[project]
name = "jarjarquant"
version = "0.1.0"
description = "Implements common financial ML techniques"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = [
    "asttokens==2.4.1",
    "beautifulsoup4==4.12.3",
    "certifi==2024.8.30",
    "charset-normalizer==3.3.2",
    "colorama==0.4.6",
    "comm==0.2.2",
    "contourpy==1.3.0",
    "cycler==0.12.1",
    "cython==3.0.12",
    "debugpy==1.8.6",
    "decorator==5.1.1",
    "eventkit==1.0.3",
    "executing==2.1.0",
    "fonttools==4.54.1",
    "frozendict==2.4.4",
    "html5lib==1.1",
    "httpx>=0.28.1",
    "ib-async==1.0.3",
    "idna==3.10",
    "iniconfig==2.0.0",
    "iprogress>=0.4",
    "ipykernel==6.29.5",
    "ipython==8.28.0",
    "ipywidgets>=8.1.7",
    "jedi==0.19.1",
    "joblib==1.4.2",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "kiwisolver==1.4.7",
    "lxml==5.3.0",
    "matplotlib==3.9.2",
    "matplotlib-inline==0.1.7",
    "multitasking==0.0.11",
    "nest-asyncio==1.6.0",
    "numpy==2.1.1",
    "packaging==24.1",
    "pandas==2.2.3",
    "parso==0.8.4",
    "patsy==1.0.1",
    "peewee==3.17.6",
    "pillow==10.4.0",
    "platformdirs==4.3.6",
    "pluggy==1.5.0",
    "polars>=1.31.0",
    "prompt-toolkit==3.0.48",
    "psutil==6.0.0",
    "pure-eval==0.2.3",
    "pyarrow>=20.0.0",
    "pygments==2.18.0",
    "pyparsing==3.1.4",
    "pytest==8.3.4",
    "pytest-asyncio==0.25.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.1.1",
    "pytz==2024.2",
    "pywin32==307 ; sys_platform == 'win32'",
    "pyzmq==26.2.0",
    "requests==2.32.3",
    "scikit-learn==1.5.2",
    "scipy==1.14.1",
    "setuptools==75.1.0",
    "six==1.16.0",
    "soupsieve==2.6",
    "stack-data==0.6.3",
    "statsmodels==0.14.4",
    "threadpoolctl==3.5.0",
    "tornado==6.4.1",
    "tqdm>=4.67.1",
    "traitlets==5.14.3",
    "tzdata==2024.2",
    "urllib3==2.2.3",
    "wcwidth==0.2.13",
    "webencodings==0.5.1",
    "yfinance==0.2.54",
]

[build-system]
requires = [
    "setuptools>=42",
    "wheel",
    "Cython",
    "numpy"
]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "ruff>=0.12.0",
]
